% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_native.R
\name{bricklayer_json_to_json}
\alias{bricklayer_json_to_json}
\title{Serialise an R object to JSON}
\usage{
bricklayer_json_to_json(x, pretty = FALSE, ...)
}
\arguments{
\item{x}{object to serialise.}

\item{pretty}{FALSE, TRUE, or an indent width.}

\item{...}{jsonlite's toJSON options: dataframe, matrix, Date, POSIXt,
factor, complex, raw, null, na, auto_unbox, digits, force.}
}
\value{
a length-one character vector of JSON.
}
\description{
Serialise an R object to JSON
}
